+++ src/components/AccountFormFields/index.tsx
@@ -53,13 +53,16 @@ function AccountFormFields({
     setShowPassword(false);
   }, []);
 
-  // Note: As of 1/22, the password must not contain a period to allow
-  // patrons to log into Overdrive for digital reading.
-  const validatePassword = (val) => {
+  const validatePasswordLength = (val) => {
     return (
-      val.length >= minPasswordLength &&
-      val.length <= maxPasswordLength &&
-      val.indexOf(".") === -1
+      (val.length >= minPasswordLength && val.length <= maxPasswordLength) ||
+      t("account.errorMessage.password")
+    );
+  };
+  const verifyPasswordmatch = () => {
+    return (
+      getValues("password") === getValues("verifyPassword") ||
+      t("account.errorMessage.verifyPassword")
     );
   };
 
@@ -84,8 +87,10 @@ function AccountFormFields({
             minLength={minPasswordLength}
             maxLength={maxPasswordLength}
             ref={register({
-              validate: (val) =>
-                validatePassword(val) || t("account.errorMessage.password"),
+              validate: {
+                validatePasswordLength,
+                verifyPasswordmatch,
+              },
             })}
             defaultValue={formValues.password}
           />
@@ -105,9 +110,7 @@ function AccountFormFields({
             minLength={minPasswordLength}
             maxLength={maxPasswordLength}
             ref={register({
-              validate: (val) =>
-                val === getValues("password") ||
-                t("account.errorMessage.verifyPassword"),
+              validate: verifyPasswordmatch,
             })}
             defaultValue={formValues.verifyPassword}
           />