+++ src/components/ApiErrors/index.tsx
@@ -30,12 +30,25 @@ const ApiErrors = React.forwardRef<HTMLDivElement, ApiErrorsProps>(
     if (!problemDetail || problemDetail?.status < 400) {
       return null;
     }
-
     if (typeof problemDetail !== "string" && !problemDetail.detail) {
       console.log("ApiErrors problemDetail: ", problemDetail);
       problemDetail.detail = t("apiErrors.defaultError");
     }
-    if (lang !== "en" && typeof problemDetail !== "string") {
+    if (
+      lang === "en" &&
+      typeof problemDetail !== "string" &&
+      problemDetail.detail.includes("PIN is trivial")
+    ) {
+      problemDetail.detail =
+        "Password cannot contain consecutively repeating characters three or more times, e.g. aaaatf54 or repeating a pattern, e.g. abcabcab";
+    }
+    if (
+      lang !== "en" &&
+      typeof problemDetail !== "string" &&
+      problemDetail.detail.includes("PIN is trivial")
+    ) {
+      problemDetail.detail = t("account.password.instruction");
+    } else if (lang !== "en" && typeof problemDetail !== "string") {
       const errorToTranslate = problemDetail?.detail;
       const newErrorMessage = errorToTranslate
         ? apiErrorTranslations[errorToTranslate][lang]